requirements.txt
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
uvicorn
fastapi
python-jose
passlib
email-validator
cornac
numpy
pandas
